# implementations/go/examples/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the DDD-based Account domain example service (Account/Card/Payment/Refund/Credential Bounded Contexts). Generated from the @-annotations above each handler in internal/interface/http/ by swag (docs/architecture/api-documentation.md) — never hand-edited.",
        "title": "Account Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/accounts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Opens a new account for the authenticated requester with a 0 balance in the given currency.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Open a new account",
                "parameters": [
                    {
                        "description": "Owner email and currency",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The account was created.",
                        "schema": {
                            "$ref": "#/definitions/http.CreateAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Request validation failed (`VALIDATION_FAILED`) — e.g. an invalid or missing email.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the account only if it belongs to the authenticated requester.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Look up an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The account was found.",
                        "schema": {
                            "$ref": "#/definitions/http.GetAccountResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No account exists with the given `id` for this requester (`ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently closes an account. The balance must be exactly 0 first (withdraw or transfer out any remaining funds).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Close an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The account was closed."
                    },
                    "400": {
                        "description": "One of: the account is already closed (`ACCOUNT_ALREADY_CLOSED`), or the balance is not 0 (`ACCOUNT_BALANCE_NOT_ZERO`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No account exists with the given `id` for this requester (`ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/deposit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Credits the given amount to the account and records a `DEPOSIT` transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Deposit money into an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Deposit amount",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The deposit succeeded.",
                        "schema": {
                            "$ref": "#/definitions/http.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "One of: request validation failed (`VALIDATION_FAILED`), the amount is not a positive integer (`ACCOUNT_INVALID_AMOUNT`), or the account is not active (`ACCOUNT_DEPOSIT_REQUIRES_ACTIVE_ACCOUNT`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No account exists with the given `id` for this requester (`ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/reactivate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Moves a suspended account back to active, restoring its ability to accept deposits/withdrawals/transfers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Reactivate a suspended account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The account was reactivated."
                    },
                    "400": {
                        "description": "Only a suspended account can be reactivated (`ACCOUNT_REACTIVATE_REQUIRES_SUSPENDED_ACCOUNT`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No account exists with the given `id` for this requester (`ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{id}/suspend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Suspends an active account, blocking further deposits/withdrawals/transfers until it is reactivated.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Suspend an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The account was suspended."
                    },
                    "400": {
                        "description": "Only an active account can be suspended (`ACCOUNT_SUSPEND_REQUIRES_ACTIVE_ACCOUNT`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No account exists with the given `id` for this requester (`ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CreateAccountRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "The account's currency code (e.g. `KRW`, `USD`). Fixed for the account's lifetime.",
                    "type": "string"
                },
                "email": {
                    "description": "The account owner's email address. Used as the destination for notification emails.",
                    "type": "string"
                }
            }
        },
        "http.CreateAccountResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "The newly created account's ID.",
                    "type": "string"
                },
                "balance": {
                    "description": "The account balance. Always 0 immediately after creation.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/http.MoneyResponse"
                        }
                    ]
                },
                "createdAt": {
                    "description": "When the account was created.",
                    "type": "string"
                },
                "email": {
                    "description": "The account owner's email address.",
                    "type": "string"
                },
                "ownerId": {
                    "description": "The authenticated requester's user ID.",
                    "type": "string"
                },
                "status": {
                    "description": "The account status (`ACTIVE`, `SUSPENDED`, or `CLOSED`).",
                    "type": "string"
                }
            }
        },
...
        "http.DepositRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "The amount to credit. Must be a positive integer.",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token issued by `POST /auth/sign-in`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}